refactor: ConfigurableListのYAGNI原則に基づく整理とRSC/RCC境界の修正
## 変更内容

### 1. YAGNI原則に基づくコンポーネント整理
- 不要な分割コンポーネント（BasicList、SortableList、FilterableList）を削除
- ConfigurableListを単一の堅牢なコンポーネントとして維持（28KB、UX優先）
- 17ファイル中15ファイルが全機能を必要としていたため、分割のメリットが僅少と判断

### 2. ファイル構造の再編成
- `custom/list/configurable-list.tsx` → `custom/configurable-list.tsx`へ移動
- フック類を`configurable-list/hooks/`に配置してプロジェクト規約に準拠
- テストとStorybookファイルを適切な位置に移動

### 3. RSC/RCC境界の修正
- React Hooksに"use client"ディレクティブを追加
- index.tsからHooksエクスポートを削除（Server Component互換性のため）
- configurable-list.client.tsを作成してClient Component専用エクスポートを提供
- RSC-RCC-BOUNDARY.mdドキュメントを作成

## 背景
- 当初ConfigurableListを3つのコンポーネントに分割を試みたが、UXが著しく劣化
- フィルターUIがドロワーになる等、使い勝手が悪化
- tree-shakingによるバンドルサイズ削減効果も限定的（28KB→最小18KB）

## 解決した問題
- "useEffect only works in Client Component"エラーの解消
- Server ComponentがReact Hooksをインポートできない問題の修正
- ビルドとランタイムエラーの解消
