+++ apps/web/src/app/circles/[circleId]/components/CircleWorksList.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import type { WorkListResultPlain, WorkPlainObject } from "@suzumina.click/shared-types";
-import { ConfigurableList } from "@suzumina.click/ui/components/custom/list";
+import { ConfigurableList } from "@suzumina.click/ui/components/custom";
 import { useCallback, useMemo } from "react";
 import { ListWrapper } from "@/components/list/ListWrapper";
 import { WorkListItem } from "@/components/work/WorkListItem";
