+++ apps/web/src/app/creators/[creatorId]/components/CreatorPageClient.tsx
@@ -6,7 +6,7 @@ import {
 	ListPageContent,
 	ListPageHeader,
 	ListPageLayout,
-} from "@suzumina.click/ui/components/custom/list-page-layout";
+} from "@suzumina.click/ui/components/custom";
 import CreatorWorksList from "./CreatorWorksList";
 
 interface CreatorPageClientProps {
