+++ apps/web/src/app/videos/page.tsx
@@ -2,7 +2,7 @@ import {
 	ListPageContent,
 	ListPageHeader,
 	ListPageLayout,
-} from "@suzumina.click/ui/components/custom/list-page-layout";
+} from "@suzumina.click/ui/components/custom";
 import { Suspense } from "react";
 import { getVideosList } from "./actions";
 import VideoList from "./components/VideoList";
