+++ apps/web/src/components/content/works-page-client.tsx
@@ -5,7 +5,7 @@ import {
 	ListPageContent,
 	ListPageHeader,
 	ListPageLayout,
-} from "@suzumina.click/ui/components/custom/list-page-layout";
+} from "@suzumina.click/ui/components/custom";
 import { Suspense } from "react";
 import WorksList from "@/app/works/components/WorksList";
 import { useAgeVerification } from "@/contexts/age-verification-context";
