+++ packages/ui/src/components/custom/configurable-list.client.ts
@@ -1,0 +1,9 @@
+/**
+ * Client Component専用のエクスポート
+ * React HooksはこちらからインポートしてClient Componentでのみ使用する
+ */
+
+"use client";
+
+export { useListData } from "./configurable-list/hooks/useListData";
+export { useListUrl } from "./configurable-list/hooks/useListUrl";
