--- packages/ui/src/components/custom/list/index.ts
@@ -1,39 +1,0 @@
-/**
- * リストコンポーネントのエクスポート
- */
-
-// 新しい分割されたコンポーネント
-export { BasicList, type BasicListProps } from "./basic-list";
-// 既存のConfigurableList（互換性のため）
-export { ConfigurableList } from "./configurable-list";
-// フック
-export { useListData } from "./core/hooks/useListData";
-export { useListUrl } from "./core/hooks/useListUrl";
-// Core types and utilities
-export type {
-	ConfigurableListProps,
-	DataAdapter,
-	FilterConfig,
-	ListController,
-	ListDataSource,
-	ListError,
-	SortConfig,
-	StandardListParams,
-} from "./core/types";
-export {
-	calculatePagination,
-	createDataAdapter,
-	wrapLegacyFetchData,
-} from "./core/utils/dataAdapter";
-export {
-	generateOptions,
-	generateYearOptions,
-	getDefaultFilterValues,
-	hasActiveFilters,
-	isFilterEnabled,
-	normalizeOptions,
-	transformFilterValue,
-	validateFilterValue,
-} from "./core/utils/filterHelpers";
-export { FilterableList, type FilterableListProps } from "./filterable-list";
-export { SortableList, type SortableListProps } from "./sortable-list";
